Suppress ImportException
When a Conda environment is destroyed, packages we depend on
in our pre-unlink scripts might not be in the env anymore.
This is a Conda problem that needs to be fixed by linking
in toposorted order and unlinking in the reverse order.
For now, we will just swallow those exceptions. The Jupyter
config file we are trying to deregister from lives in the env
and will be destroyed anyway.

For cases where we're just removing an extention, the dependencies
will still be installed properly and we'll unregister from the
config files properly.